move class DownloadedFile to the proper package
