+++ src/main/java/com/codeborne/selenide/files/DownloadedFile.java
@@ -1,4 +1,4 @@
-package com.codeborne.selenide.proxy;
+package com.codeborne.selenide.files;
 
 import javax.annotation.CheckReturnValue;
 import javax.annotation.Nonnull;
