+++ src/main/java/com/codeborne/selenide/files/FileFilter.java
@@ -1,7 +1,5 @@
 package com.codeborne.selenide.files;
 
-import com.codeborne.selenide.proxy.DownloadedFile;
-
 import java.io.Serializable;
 
 public interface FileFilter extends Serializable {
