+++ src/main/java/com/codeborne/selenide/files/FilenameFilter.java
@@ -1,7 +1,5 @@
 package com.codeborne.selenide.files;
 
-import com.codeborne.selenide.proxy.DownloadedFile;
-
 class FilenameFilter implements FileFilter {
   private final String fileName;
