+++ src/main/java/com/codeborne/selenide/impl/Downloads.java
@@ -1,7 +1,7 @@
 package com.codeborne.selenide.impl;
 
 import com.codeborne.selenide.files.FileFilter;
-import com.codeborne.selenide.proxy.DownloadedFile;
+import com.codeborne.selenide.files.DownloadedFile;
 
 import javax.annotation.CheckReturnValue;
 import javax.annotation.Nonnull;
