+++ src/test/java/com/codeborne/selenide/impl/DownloadDetectorTest.java
@@ -1,6 +1,6 @@
 package com.codeborne.selenide.impl;
 
-import com.codeborne.selenide.proxy.DownloadedFile;
+import com.codeborne.selenide.files.DownloadedFile;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
